Refactor agGrid Tables to use a mixin
